% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pDC}
\alias{pDC}
\title{Sample pDC cells from Covid PBMC data}
\format{
A Seurat object
}
\source{
\url{https://doi.org/10.1038/s41591-021-01329-2}
}
\usage{
data(pDC)
}
\description{
A dataset containing 100 randomly sampled pDC cells for
demonstrating CellDEEP functions.
}
