% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_all_function.R
\name{pseudobulk_dge_pool}
\alias{pseudobulk_dge_pool}
\title{Pseudobulk Function}
\usage{
pseudobulk_dge_pool(sce, contrasts, verbose = FALSE, ...)
}
\arguments{
\item{sce}{The SingleCellExperiment object want to be processed}

\item{contrasts}{The contrast matrix used to indicate comparison user want to make}

\item{verbose}{Show the input comparison matrix or not}

\item{...}{Places for other further possible parameter}
}
\description{
This function pseudobulk the input sce object, and run DE analysis as contrast indicate
}
