% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Collection_of_poolingFunctions.R
\name{random.cellPooling.dev.yiyi.2}
\alias{random.cellPooling.dev.yiyi.2}
\title{randomly pooling cells together, if sample to pool has less than 25 cells, jump it}
\usage{
random.cellPooling.dev.yiyi.2(
  dataset,
  n_cells = 10,
  assay_name = "RNA",
  readcounts = "mean"
)
}
\arguments{
\item{dataset}{A seurat object}

\item{n_cells}{Number of cells to pool together}

\item{assay_name}{Which assay to work on}

\item{readcounts}{"sum" or "mean".}
}
\description{
randomly pooling cells together, if sample to pool has less than 25 cells, jump it
}
