% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_all_function.R
\name{random.cellPooling.dev.yiyi}
\alias{random.cellPooling.dev.yiyi}
\title{Pooling cells randomly.}
\usage{
random.cellPooling.dev.yiyi(
  dataset,
  n_cells = 10,
  assay_name = "RNA",
  readcounts = "mean",
  cell_cutoff = 25
)
}
\arguments{
\item{dataset}{A seurat object}

\item{n_cells}{The number of cells to pool together.}

\item{assay_name}{The assay to pool based_on}

\item{readcounts}{"sum" or "mean". How to generate read counts when pooling cells together.}

\item{cell_cutoff}{Cell filtering number, After subset according to sample, group and cluster,if remaining cells lower than this number, these cells will be filtered out.Default 25.}
}
\description{
This function pool cells randomly.
}
