% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Collection_of_poolingFunctions.R
\name{replicate.splitting}
\alias{replicate.splitting}
\title{create 2 fake replicates for pesudobulk sample}
\usage{
replicate.splitting(
  dataset,
  sample_to_split = "ZC3H20_KO",
  subid1 = "ZC3H20_KO_1",
  subid2 = "ZC3H20_KO_2"
)
}
\arguments{
\item{dataset}{The input Seurat object.}

\item{sample_to_split}{The sample_id to split}

\item{subid1}{The 1st sub sample_id}

\item{subid2}{The 2nd sub sample_id}
}
\description{
create 2 fake replicates for pesudobulk sample
}
