% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_FindMarker.R
\name{return.DE}
\alias{return.DE}
\title{The function to return DE genes}
\usage{
return.DE(
  dataset,
  test.use = "wilcox",
  DE.ident.1,
  DE.ident.2,
  DE.group,
  assay = "RNA",
  p_cutoff = 0.05
)
}
\arguments{
\item{dataset}{A seurat object}

\item{test.use}{method to run DE analysis.eg: "MAST", "DESeq2"....(check in seurat FindMarker command)}

\item{assay}{The assay to run analysis, "RNA" as default}

\item{p_cutoff}{P-value cutoff for DE genes, 0.05 as default}

\item{ident.1}{The 1st ident to compare, can be missing}

\item{ident.2}{The 2nd ident to comapre, can be missing}

\item{group.by}{The group to run DE analysis}
}
\description{
The function to return DE genes
}
