% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{data}
\name{sample}
\alias{sample}
\title{Sample data as example
A dataset contains 100 randomly sample from pDC cells in Covid PBMC data.}
\source{
\url{https://doi.org/10.1038/s41591-021-01329-2}
}
\description{
Sample data as example
A dataset contains 100 randomly sample from pDC cells in Covid PBMC data.
}
