% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Collection_of_poolingFunctions.R
\name{cellPooling.kmean.dev.yiyi}
\alias{cellPooling.kmean.dev.yiyi}
\title{Pooling cells by k-mean clustering.}
\usage{
cellPooling.kmean.dev.yiyi(
  dataset,
  n_cells = 10,
  nstart = 100,
  assay_name = "RNA",
  readcounts = "mean",
  cell_cutoff = 25
)
}
\arguments{
\item{dataset}{A Seurat object}

\item{n_cells}{number of cells to pool together}

\item{nstart}{the nstart in kmeans clustering, which represents how many sets to start with.}

\item{assay_name}{the assay to pool}

\item{readcounts}{"mean" or "sum" or "10X". How to treat read counts when pool it. "mean" is mean the read counts and round; "sum" is sum; "10X" is the mean of read counts and 10 times it}
}
\description{
This function pool cells by k-mean clustering.
}
