% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare.R
\name{check.ground.truth}
\alias{check.ground.truth}
\title{check a ground truth (muscat) table vs a list of genes DE}
\usage{
check.ground.truth(ground.truth.table, genes.DE, title, verbose = FALSE)
}
\arguments{
\item{ground.truth.table}{table with ground truth genes}

\item{genes.DE}{The list of DE genes that user want to check}

\item{title}{The title to put in output, eg: "MAST"}

\item{verbose}{Show the info of ground truth or not}
}
\value{
ground truth table VS the DE gene list
}
\description{
This function take a seurat object as input, and subset it as user input.
}
\examples{
\donttest{
#Return DE
markers.k.pooled.sum.15cell.MAST.DE <- return.DE(k.pooled.sum.15cell, "MAST")
markers.k.pooled.sum.15cell.DESEQ.DE <- return.DE(k.pooled.sum.15cell, "DESeq2")
#Check ground truth
check.ground.truth(Ground.truth, markers.k.pooled.sum.15cell.MAST.DE, title = "kmeans (sum) + MAST")
check.ground.truth(Ground.truth, markers.k.pooled.sum.15cell.DESEQ.DE, title = "kmeans (sum) + DESeq2")
}

}
