% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_all_function.R
\name{FindMarker.CellDEEP}
\alias{FindMarker.CellDEEP}
\title{Funciton to run CellDEEP with DE analysis}
\usage{
FindMarker.CellDEEP(
  object,
  ident.1 = NULL,
  ident.2 = NULL,
  group.by = NULL,
  test.use = "wilcox",
  Pool = TRUE,
  readcounts = "sum",
  n_cells = 10,
  assay = "RNA",
  cell_cutoff = 25,
  pool_way = "kmean",
  pvalue_cutoff = 0.05,
  ...
)
}
\arguments{
\item{object}{The input seurat object (cells user want to pool)}

\item{ident.1}{The first ident to compare for DE}

\item{ident.2}{The second ident to compare for DE}

\item{group.by}{The group comparison for DE (metadata name where the DE will be applied)}

\item{test.use}{The DE method to use}

\item{Pool}{Pool cells or not, default TRUE, when user want to run DE withour any pooling set to FASLE}

\item{readcounts}{The way to treat readcounts while pooling, options: "sum" and "mean" (For Kmean clustering pooling, one extra option "10X", which is multiple mean readcounts 10 times )}

\item{n_cells}{How many cells to pool together, this parameter affects method performance, check paper for details. 10 as defatult.}

\item{assay}{The assay to process. "RNA" as default.}

\item{cell_cutoff}{Cell filtering number, After subset according to sample, group and cluster,if remaining cells lower than this number, these cells will be filtered out.Default 25.}

\item{pool_way}{The pooling atrategy, Two option: "kmean" and "random".}

\item{pvalue_cutoff}{The adjusted p-value cutoff for DE analysis.}

\item{...}{Places for other further possible parameter}
}
\description{
This function run CellDEEP, pooling cells then run DE analysis
}
