% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_FindMarker.R
\name{FindMarker.CellDEEP}
\alias{FindMarker.CellDEEP}
\title{FindMarker with CellDEEP}
\usage{
FindMarker.CellDEEP(
  object,
  ident.1 = NULL,
  ident.2 = NULL,
  group.by = NULL,
  test.use = "wilcox",
  Pool = FALSE,
  readcounts = "sum",
  n_cells = 10,
  assay = "RNA",
  cell_cutoff = 25,
  ...
)
}
\arguments{
\item{object}{A Seurat Object}

\item{ident.1}{1st ident to compare}

\item{ident.2}{2nd ident to compare}

\item{group.by}{the group to compare}

\item{...}{}
}
\description{
FindMarker with CellDEEP
}
