% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare.R
\name{prepare_data}
\alias{prepare_data}
\title{Adjust input data format, Normalize/Scale/UMAP if needed, make the object ready for further steps.}
\usage{
prepare_data(
  obj,
  assay = "RNA",
  sample_id,
  group_id,
  cluster_id,
  Ident.to.subset = NULL,
  Need.to.Norm_Transf = FALSE,
  Need.to.Scale = FALSE,
  Need.UMAP.generation = FALSE,
  Ground.truth = FALSE,
  Specific.Idents.to.subset = NULL,
  file_path = NULL
)
}
\arguments{
\item{obj}{A Seurat object, should have metadata for sample, cluster and group info.}

\item{assay}{RNA as default. If assay = "covid", create a new seurat obejct based on RAW read counts of input
data, and go with it.}

\item{sample_id}{The metadata has sample information.}

\item{group_id}{The metadata has group information.}

\item{cluster_id}{The metadata has cluster information.}

\item{Ident.to.subset}{The idents to subset(if user want to subset the data).}

\item{Need.to.Norm_Transf}{If the object need normalizarion or not.}

\item{Need.to.Scale}{If the object need scale or not}

\item{Need.UMAP.generation}{If the object need generate an UMAP or not.}

\item{Ground.truth}{Does this dataset have ground truth or not.}

\item{Specific.Idents.to.subset}{If need subset, which element do you want to choose.}

\item{file_path}{The file path to store ground truth}
}
\description{
This function adjust input data format to make it ready for further steps. Add "sample_id",
"cluster_id" and "group_id". Can Normalize/Scale/RunUMAP if needed.
}
