% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Collection_of_poolingFunctions.R
\name{random.cellPooling.dev.yiyi}
\alias{random.cellPooling.dev.yiyi}
\title{Pooling cells randomly.}
\usage{
random.cellPooling.dev.yiyi(
  dataset,
  n_cells = 10,
  assay_name = "RNA",
  readcounts = "mean"
)
}
\arguments{
\item{dataset}{A seurat object}

\item{n_cells}{The number of cells to pool together.}

\item{assay_name}{The assay to pool based_on}

\item{readcounts}{"sum" or "mean". How to generate read counts when pooling cells together.}
}
\description{
This function pool cells randomly.
}
