% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_all_function.R
\name{return.DE}
\alias{return.DE}
\title{Function to run DE analysis: FindMakrer + P-value filtering}
\usage{
return.DE(
  dataset,
  test.use = "wilcox",
  DE.ident.1,
  DE.ident.2,
  DE.group,
  assay = "RNA",
  p_cutoff = 0.05
)
}
\arguments{
\item{dataset}{Input seurat object}

\item{test.use}{The DE analysis method}

\item{DE.ident.1}{The first ident to compare}

\item{DE.ident.2}{The second ident to compare}

\item{DE.group}{The metadata where comparison based on - eg: group_id}

\item{assay}{The assay to run DE, default "RNA"}

\item{p_cutoff}{The adjusted p_value cutoff, 0.05 as default}
}
\description{
This function perform DE analysis with FindMarker
}
