% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim}
\alias{sim}
\title{Sample simulated cells from muscat package}
\format{
A Seurat object
}
\source{
simulated data with muscat package
}
\usage{
data(sim)
}
\description{
A dataset containing 200 simulated cells(100 per group) for
demonstrating CellDEEP functions.
Can be found at \doi{10.5281/zenodo.18863779}
}
