% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_all_function.R
\name{Subset.seurat.object}
\alias{Subset.seurat.object}
\title{Subset Seurat object}
\usage{
Subset.seurat.object(Seurat.object, Ident.to.subset, Specific.Idents.to.subset)
}
\arguments{
\item{Seurat.object}{A Seurat object user want to subset}

\item{Ident.to.subset}{The ident to subset, eg: cluster_id}

\item{Specific.Idents.to.subset}{The ident want to susbet, eg: 5}
}
\value{
Subset Seurat object
}
\description{
This function take a seurat object as input, and subset it as user input.
}
